Refactor test framework for scalability and developer ease
BREAKING CHANGE: Mock files renamed from *_mock.h to match real headers

Changes:
- Renamed all mock files to match actual header names (logger_mock.h -> logger.h)
- Removed wrapper headers from test directories - no longer needed
- Created samwise_add_test() CMake helper function for easy test creation
- Mock headers now shadow real headers via include path ordering
- Added comprehensive TESTING.md documentation
- CTest now properly discovers and runs tests

Developer Benefits:
- Write tests with normal #includes - no special mock headers
- Simple 3-line CMake configuration per test using samwise_add_test()
- CMake handles all header routing automatically via include paths
- Tests now work with ctest for automated test discovery and reporting

Example test CMakeLists.txt:
  samwise_add_test(
    NAME my_test
    SOURCES my_test.c
    LIBRARIES my_task_lib
  )

Verified:
- Tests build successfully: cmake -B build_tests -DPROFILE=TESTS
- Tests run with ctest: 100% tests passed (1/1)
- Embedded builds unchanged: cmake -B build_embedded -DPROFILE=PICUBED_DEBUG
