+++ src/tasks/print/test/CMakeLists.txt
@@ -1,14 +1,7 @@
-add_executable(print_task_test print_test.c)
-
-target_link_libraries(print_task_test PRIVATE
-  print_task
-  test_mocks
+# Example test using the samwise_add_test helper
+# Developers only need to specify the test name, source files, and libraries to test
+samwise_add_test(
+  NAME print_task_test
+  SOURCES print_test.c
+  LIBRARIES print_task
 )
-
-target_include_directories(print_task_test PRIVATE
-  "${CMAKE_CURRENT_SOURCE_DIR}"
-  "${PROJECT_SOURCE_DIR}/src/test_mocks"
-  "${PROJECT_SOURCE_DIR}/src/tasks/print"
-)
-
-add_test(NAME PrintTaskTest COMMAND print_task_test)
