--- src/tasks/print/test/error.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "error_mock.h"
