--- src/tasks/print/test/logger.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "logger_mock.h"
