--- src/tasks/print/test/macros.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "macros_mock.h"
