--- src/tasks/print/test/neopixel.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "neopixel_mock.h"
