--- src/tasks/print/test/slate.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "slate_mock.h"
