--- src/tasks/print/test/state_machine.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "state_machine_mock.h"
