--- src/tasks/print/test/typedefs.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "typedefs_mock.h"
