+++ src/test_mocks/macros.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "error_mock.h"
-#include "logger_mock.h"
+#include "error.h"
+#include "logger.h"
 
 #ifdef BRINGUP
 #define IS_BRINGUP true
