+++ src/test_mocks/neopixel.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-#include "neopixel_mock.h"
+#include "neopixel.h"
 
 void neopixel_set_color_rgb(uint8_t r, uint8_t g, uint8_t b)
 {
