+++ src/test_mocks/slate.h
